+++ Nos/Service/FileStorage/URLSessionProtocol.swift
@@ -8,12 +8,14 @@ extension URLSession: URLSessionProtocol {}
 
 class MockURLSession: URLSessionProtocol {
     private let responseData: Data
+    private let urlResponse: URLResponse
 
-    init(responseData: Data = Data()) {
+    init(responseData: Data = Data(), urlResponse: URLResponse = URLResponse()) {
         self.responseData = responseData
+        self.urlResponse = urlResponse
     }
 
     func data(for request: URLRequest) async throws -> (Data, URLResponse) {
-        (responseData, URLResponse())
+        (responseData, urlResponse)
     }
 }